requirements.txt
pandas
numpy
pillow
opencv-python
matplotlib
jupyter
ipykernel
pytesseract
python-dotenv
tqdm
playwright
pytest
nbformat

openai>=1.40.0

# Playwright note: after installing, run `playwright install` in your environment.

openpyxl
